Day 17 parts 1 and 2
Probably could do with a refactor, I mean the finding
code is similar, however, I don't really think it's
worth it. I mean in the one case you want the shortest,
in the other, the longest, so they really are quite
different things.

I went down the route of adding callbacks etc, but it's
definitely not any clearer.