day 23 part 2 complete
I gave up on this one, :( maybe I could/should have worked it out.
I could see the loops but my brain was hurting too much. So I looked at
the reddit ->
https://www.reddit.com/r/adventofcode/comments/7lo2ed/can_someone_explain_day_23_part_2/
and then stepped through that working it out.

Certainly indenting the loops is a good way to start.